// src/types/cssProps.ts
export type BadgeCSSProps = {
    badgeClassName?: string;
    badgeStyle?: React.CSSProperties;

    badgeContainerClassName?: string;
    badgeContainerStyle?: React.CSSProperties;

    badgeContainerLightClassName?: string;
    badgeContainerLightStyle?: React.CSSProperties;

    badgeContainerDarkClassName?: string;
    badgeContainerDarkStyle?: React.CSSProperties;

    badgeGoogleIconClassName?: string;
    badgeGoogleIconStyle?: React.CSSProperties;

    badgeInnerContainerClassName?: string;
    badgeInnerContainerStyle?: React.CSSProperties;

    badgeLabelClassName?: string;
    badgeLabelStyle?: React.CSSProperties;

    badgeLabelLightClassName?: string;
    badgeLabelLightStyle?: React.CSSProperties;

    badgeLabelDarkClassName?: string;
    badgeLabelDarkStyle?: React.CSSProperties;

    badgeRatingContainerClassName?: string;
    badgeRatingContainerStyle?: React.CSSProperties;

    badgeRatingClassName?: string;
    badgeRatingStyle?: React.CSSProperties;

    badgeRatingLightClassName?: string;
    badgeRatingLightStyle?: React.CSSProperties;

    badgeRatingDarkClassName?: string;
    badgeRatingDarkStyle?: React.CSSProperties;

    badgeStarsClassName?: string;
    badgeStarsStyle?: React.CSSProperties;

    badgeStarsContainerClassName?: string;
    badgeStarsContainerStyle?: React.CSSProperties;

    badgeStarsFilledClassName?: string;
    badgeStarsFilledStyle?: React.CSSProperties;

    badgeStarsEmptyClassName?: string;
    badgeStarsEmptyStyle?: React.CSSProperties;

    badgeLinkContainerClassName?: string;
    badgeLinkContainerStyle?: React.CSSProperties;

    badgeLinkClassName?: string;
    badgeLinkStyle?: React.CSSProperties;

    badgeLinkLightClassName?: string;
    badgeLinkLightStyle?: React.CSSProperties;

    badgeLinkDarkClassName?: string;
    badgeLinkDarkStyle?: React.CSSProperties;

    badgeLinkInlineClassName?: string;
    badgeLinkInlineStyle?: React.CSSProperties;
};